split genic into libgenic.
