+++ libgenic/allvars.h
