+++ libgenic/power.h
