+++ libgenic/thermal.c
