+++ libgenic/thermal.h
